moves get_attribute_doc to helpers
